prepare nettle-sys
Note: we have manually added nettle-dev to the list of
Build-Dependencies and Dependencies by modifying debian/control.  It's
needed for the Build-Dependencies to make sure that it builds
correctly, and the -dev package also requires it, because anything
using the -dev package is going to fail if nettle-dev isn't also
installed.