Update pretty_assertions to 0.6.1-2: fix tests, relax deps
With newer Rust versions, the test suite for pretty_assertions fails due
to formatting changes, see [0].  As this has not been fixed in upstream
for more than six months, I manually patched the tests.  As ansi_term
has been updated in the meantime, we also have to relax the ansi_term
dependency to work with 0.12.

[0] https://github.com/colin-kiegel/rust-pretty-assertions/issues/32