log: prepare upload ignoring sval features
In trying to resolve #948639, f_g identified three different options,
since sval isn't in debian yet:

 - patch out all sval-related features
 - leave all sval-related features enabled, but make them "empty"
 - wait for sval to clear NEW

The middle option would mean that we could avoid rust-log having to
clear NEW in the future once sval is in the archive, but it would also
mean having a package in the archive (at least temporarily) that is
"false" in the sense that it claims to be librust-log+sval-dev, but
doesn't actually provide the log+sval feature.

And, sval was recently rejected from NEW. :(

So given that:

 - i don't want to put false packages in the archive, and

 - The rest of the debian rust ecosystem that *is* waitin gon a
   rust-log migration isn't depending on either librust-log+sval-dev
   or librust-log+kv_unstable-sval-dev, and

 - i have no way of predicting when sval will clear NEW

I'm going with the first option.