nettle-sys: bump bindgen to 0.53.1, up to 0.55.x
bindgen 0.53.0 introduced a change in how size_t is mapped:

https://github.com/rust-lang/rust-bindgen/issues/1671

This change would produce an API change in any bindgen-created
modules, and is not relevant on any of the platforms that we target
(maybe not on any platform in existence).

bindgen 0.53.1 introduced size_t_is_usize() to revert to the old
behavior, which should keep the nettle-sys API stable.

See also https://gitlab.com/sequoia-pgp/nettle-sys/-/merge_requests/7
where i'm trying to get this same change adopted upstream.