build.sh: reset the append-only option lists
sbuild arguments like --autopkgtest-opt (and -opts) append the specified
options to associated config file list (in this case autopkgtest_opts).
This can lead to a broken configuration. For example in ~/.sbuildrc I have:

$autopkgtest_opts = ['--', 'schroot', '%r-%a-sbuild'];

and I can't use build.sh as is as it passes (basically) the same
options to sbuild via --autopkgtest-opts, and the resulting option
list with duplicate entries doesn't work.

This change ensures that the append-only option lists build.sh
uses are reset to empty lists.