Fix build of ppv-lite86 on s390x
On s390x, rustc seems to miscompile rotate_right on u128 in debug mode.

This is used in a testcase, which makes the test fail.

Fortunately, it's only used for the expected value in an assert_eq!, so
it's not really an error in the library.

By putting the value into a const, the compiler calculates the value
correctly, fixing the test case.