sequoia-openpgp: clarify tests
sequoia-openpgp requires exactly one of the crypto-* features to be enabled.

On debian, crypto-nettle is the only feasible option.

So, tests that include the crypto-nettle features should succeed.  Tests that
do not include the crypto-nettle feature are likely to fail.

We should be able to drop the librust-nettle-7+default-dev dependency because that will
be included automatically in those features that do depend on nettle.