Revert "prepare packaging of rustls-native-certs"
This reverts commit e0ffda9a540e58b95fff9f25c5b1f96388bcad77.

Reason: I packaged rustls-native-certs 0.5.0 which depends on
rustls >=0.19.0. As that is not available in debian, I changed the
dependency to >=0.18.0, which seemed to work fine.

But later I noticed that this may not be expected by users of
the package: After all, the relase 0.4.0 of rustls-native-certs
was compatible with rustls 0.18.0. So a reasonable developer
would expect to use 0.4.0 with 0.18.0 or 0.5.0 with 0.19.0.

There is no good reason for debian to deviate from the upstream
dependencies. At least not in this case.

So, for now, I retracted the upload. To avoid confusion, I also revert
the commit to debcargo-conf, although parts of this effort can
of course be reused when packaging 0.4.0 (or 0.5.0 after rustls
was upgraded.)