sequoia-openpgp: more test adjustments
upstream explicitly wants the feature named "crypto-nettle", even
though it is synonymous with the "nettle" feature.  The goal is to
require exactly one "crypto-*" feature to be enabled during a build,
so even though the features are identical functionally, they don't
have the same semantics during the build because of the feature-check.