fd-lock: import 2.0.0
This is the last of the 2.0.0 series; upstream has moved to 3.x, which
adds a bunch of changes (including a dependency on rustix, which is
more complex than i'm prepared to package right now).

I'm packaging this for a dependency that only wants fd-lock 2.0.0
anyway (pgp-cert-d), so i'm putting 2.0.0 as the initial version for
debian for now.