sequoia-octopus-librnp: initial packaging for 1.4.1
Shipping a cdylib crate in debian is fairly ugly at the moment.

See https://bugs.debian.org/1017681 and
https://bugs.debian.org/1017682 for why we can't just use debcargo and
dh-cargo more cleanly. My hope is that we can revert to using debcargo
and dh-cargo when those bugs are fixed.

We use a symlink to install the .so where thunderbird will find it
(see that package's source
comm/mail/extensions/openpgp/content/modules/RNPLib.jsm for how it
searches for librnp at runtime).