Updating scriptisto to 1.0.0 and updating/packaging its dependencies.
The strategy I've selected here is to package all that is necessary
to build the binary. Most of the deps are not used by anything else.
serde_yaml was used and it was best to just downgrade the dep
from 0.9 to 0.8 as it makes no difference to the binary, but
there are other usages of it in Debian