fasteval: fix pointer-based tests
We can't assume the pointers representation will always be as long as
the architecture's pointer width: fasteval doesn't prepend 0's for
smaller values, so it vastly depends on what pointer values are used in
userspace applications, which we can't control or anticipate.

In order to fix this, use fixed-width (8 chars) pointer representation in
all corresponding tests.