droid-juicer: update to 0.2.0
The latest upstream version introduces the possibility of a new, global
config file used for post-processing commands. This version also allows
fetching firmware from Android dynamic partitions if the system includes
`make-dynpart-mappings`.

This commit adds the latter package as a `Recommends` and ships a new
`config.toml` file for rebuilding the initramfs and replacing it in the
boot image on qcom-based devices (introducing `qcom-phone-utils` as a
new dependency to this package.