speakersafetyd: Install companion files (service, udev rules, etc)
We also want the systemd service, the udev rules, the conf files and the
blackbox directory.

To accomplish this we hook up "make install" as upstream ships a
Makefile that does this all for us.

Unfortunately the upstream Makefile also tries to install the
speakersafetyd binary which is located in a different path when we build
with the debian toolchain.
We thus temporarily make the binary available in the path the Makefile
expects it at.

For extra safety we also override the UNITDIR and UDEVDIR variables
used in the Makefile (defaulting to hardcoded values) with the value
retrieved from pkg-config files (.pc) for systemd and udev
respectively.