Work around Makefile typo for blackbox directory
The upstream speakersafetyd.service file says
/var/lib/speakersafetyd/blackbox but the Makefile
says /var/speakersafetyd/blackbox (lib/ part left out).
Override the VARDIR variable with /var/lib for now to
have the Makefile create the directory in the correct place.

This could be reverted once we update to a new upstream release
which includes a fix for:
https://github.com/AsahiLinux/speakersafetyd/issues/3