lmdb-sys: update and patch for heed
`heed` is needed as a dependency of Mousai, a GNOME Circle app. However,
it relies on `lmdb-master-sys`, which is essentially a more recent
version of `lmdb-sys`. In order to avoid carrying multiple nearly
identical crates (especially bindings to the same underlying C library),
this patch adds the following changes to `lmdb-sys`:
* add new constant and type declarations
* make all existing structs implement the `Clone` and `Copy` traits
* don't require function pointers to be `*mut`; this can be done as no
  reverse dependency uses the modified functions so far, `heed` will be
  the first user of those

Those changes are enough for `heed` to work fine with this crate instead
of `lmdb-master-sys`.