hafas-rs: fix autopkgtest failures
The profile tests rely on the local crate::requester::hyper module,
which only exists when the "hyper-requester" feature is enabled.
Therefore, those tests will fail for all single-feature runs except one.
Ensure this test is only enabled when the required feature is.

This change resulting on a lot of test cases running with 0 actual tests
being executed, override the autopkgtest control file so we only run
meaningful tests.

Closes: #1056029