sys-mount: update to 3.0.1
This allows us to drop previous patch, but requires a new one as
upstream switched to using `loopdev-3` instead of the (abandoned)
`loopdev`. However, the new crate doesn't build properly on Debian yet
(or rather, autopkgtests fail for no obvious reason), so this change is
reverted until `loopdev-3` is uploaded to Debian.