sequoia-sop: update tests for 0.32
upstream 0.32 punts the question of which dependencies of
sequoia-openpgp to use, making the default be sequoia-openpgp/default.
This way, if the builder wants something other than
sequoia-openpgp/default, the builder can specify
"--no-default-features" and then some other flavor of sequoia-openpgp.

So our testing will by default fail when --no-default-features are
supplied, so we just skip those tests.