rsop: import 0.3.8
We package both rsop and the rsopv verification-only variant.

We add an "alternatives" for sopv as well, so that verification-only
tools can find either rsop or rsopv as sopv, if they're the only thing
installed.  The comparatively low weight for the alternatives is based
on rsop's upstream's statement that rsop remains experimental.