libphosh-sys: update build dependencies
The build dependencies we need are now pulled in by `libphosh-0.x-dev`,
so we just need to mention the latter (which has been renamed for the
0.43 release, so update this as well).

The GIR file is also now part of a separate package, so let's
build-depend on it.

Finally, `dev-dependencies` are now also required as B-D, although
`debcargo` doesn't add them automatically, so add those too for the
time being.