virtiofsd: fix autopkgtest in debci
Depend on the new version of mkosi, which pulls in the required
dependencies itself.

Also restrict the test to isolation-machine, as it turns out in debci
the vhost socket is not accessible in isolation-container, so the
test cannot work.

Also skip the test if the dependencies are not available (due to versioning,
in case of backports), and if the kernel is too old, as mkosi requires
a kernel API that was added in 5.12.

The test also needs to download package, so add needs-internet.

Follow-up for 809388ca4c3c47690be17a4f880b5d477c2f8ffb