just: ignore SIGPIPE for choosers, fixes #1078430
I used to believe those tests failed due to Rust ignoring SIGPIPE by
default, thus wrote use-default-sigpipe-handler.patch which reverts to
the default handler of SIGPIPE.

After a second investigation, I now believe that is in the wrong
direction. The default handler (SIG_DFL) terminates the process, the
same when "Rust ignoring SIGPIPE". (I don't really know why the process
is terminated even then, but didn't bother. For the record, SIG_IGN also
terminates the process, IDK why either.)

SIGPIPE signals that the receiving end (in this case the chooser child
process) closed all fds to the pipe, i.e. they don't need more data from
the main process. Thus we just carry on.

As for the reason these tests only failed on s390x and ppc64el, my guess
is that they are slower, esp. in the relevant handling of signals and/or
pipes.