rustix: adapt vdso patch for old linux-raw-sys
the DT_GNU_HASH constant was only introduced in linux-raw-sys 0.6.6, this leads
to the following warning when it is not defined and vdso is attempted to be
initialized:

 warning: unreachable pattern
    --> src/backend/linux_raw/vdso.rs:183:17
     |
 178 |                 DT_GNU_HASH => {
     |                 ----------- matches any value

this only broke on i386 in a visible fashion (because of a pointer alignment
check that fails there), but is wrong on all targets.

work around it for now by hard-coding the constant (it's target-independent in
linxu-raw-sys). the workaround can be dropped once linux-raw-sys is updated to
a version providing the named constant.

Signed-off-by: Fabian Grünbichler <git@fabian.gruenbichler.email>