Defuse rustup by telling it to use the system toolchain
If rustup happens to be installed on the system where debcargo-conf
scripts are used, cargo will fail when invoked by dh-cargo.
The reason is that debhelper sets HOME to a directory inside debian/,
while dh-cargo will call the cargo wrapper which will call
/usr/bin/cargo which will run rustc from the PATH. The rustc it will run
is, however, the rustup wrapper, which will complain as it can’t find
its configuration and determine the default toolchain:

    dh clean --buildsystem cargo
       dh_auto_clean -O--buildsystem=cargo
    debian cargo wrapper: options = ['parallel=22'], profiles = [], parallel = ['-j22'], lto =
    debian cargo wrapper: rust_type = x86_64-unknown-linux-gnu, gnu_type = x86_64-linux-gnu
    debian cargo wrapper: running subprocess (['env', 'RUST_BACKTRACE=1', '/usr/bin/cargo', 'clean', '--verbose', '--verbose'],) {'check': True}
    error: process didn't exit successfully: `rustc -vV` (exit status: 1)
    --- stderr
    error: rustup could not choose a version of rustc to run, because one wasn't specified explicitly, and no default is configured.
    help: run 'rustup default stable' to download the latest stable release of Rust and set it as your default toolchain.

To work around this, provide a rust-toolchain.toml file telling
rustup to just use the system toolchain in /usr.

This file is a no-op for anyone who doesn’t use rustup.